pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ontrack-cli"
version = "0.1.0"
description = "Terminal-first CLI for the OnTrack/Doubtfire teaching platform"
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "bunizao" },
]
dependencies = [
    "browser-cookie3>=0.20",
    "click>=8.1",
    "PyYAML>=6.0",
    "requests>=2.31",
    "rich>=13.7",
]

[project.scripts]
ontrack = "ontrack_cli.cli:main"

[project.urls]
Homepage = "https://github.com/bunizao/ontrack-cli"
Repository = "https://github.com/bunizao/ontrack-cli"
Issues = "https://github.com/bunizao/ontrack-cli/issues"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.hatch.build.targets.wheel]
packages = ["ontrack_cli"]
